[ci] refactor
