+++ apps/scan/src/main/scala/org/lfdecentralizedtrust/splice/scan/store/bulk/AcsSnapshotBulkStorage.scala
@@ -7,64 +7,68 @@ import com.digitalasset.canton.data.CantonTimestamp
 import com.digitalasset.canton.logging.{NamedLoggerFactory, NamedLogging}
 import com.digitalasset.canton.time.Clock
 import com.digitalasset.canton.tracing.{Spanning, TraceContext}
-import io.grpc.Status
 import io.opentelemetry.api.trace.Tracer
 import org.apache.pekko.NotUsed
 import org.apache.pekko.actor.{ActorSystem, Cancellable}
-import org.apache.pekko.stream.scaladsl.{Sink, Source}
+import org.apache.pekko.stream.scaladsl.{Flow, Sink, Source}
 import org.apache.pekko.pattern.after
 import org.lfdecentralizedtrust.splice.PekkoRetryingService
 import org.lfdecentralizedtrust.splice.config.AutomationConfig
 import org.lfdecentralizedtrust.splice.environment.RetryProvider
-import org.lfdecentralizedtrust.splice.scan.config.{BulkStorageConfig, ScanStorageConfig}
-import org.lfdecentralizedtrust.splice.scan.store.bulk.AcsSnapshotBulkStorage.AcsSnapshotObjects
-import org.lfdecentralizedtrust.splice.scan.store.{AcsSnapshotStore, ScanKeyValueProvider}
+import org.lfdecentralizedtrust.splice.scan.config.BulkStorageConfig
+import org.lfdecentralizedtrust.splice.scan.store.AcsSnapshotStore
 import org.lfdecentralizedtrust.splice.store.S3BucketConnection.ObjectKeyAndChecksum
-import org.lfdecentralizedtrust.splice.store.{
-  HardLimit,
-  HistoryMetrics,
-  Limit,
-  S3BucketConnection,
-  TimestampWithMigrationId,
-  UpdateHistory,
-}
+import org.lfdecentralizedtrust.splice.store.{TimestampWithMigrationId, UpdateHistory}
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.concurrent.duration.*
 
-class AcsSnapshotBulkStorage(
-    storageConfig: ScanStorageConfig,
+abstract class AcsSnapshotBulkStorage(
     appConfig: BulkStorageConfig,
     acsSnapshotStore: AcsSnapshotStore,
     updateHistory: UpdateHistory,
-    s3Connection: S3BucketConnection,
-    kvProvider: ScanKeyValueProvider,
-    historyMetrics: HistoryMetrics,
     override val loggerFactory: NamedLoggerFactory,
 )(implicit actorSystem: ActorSystem, ec: ExecutionContext)
     extends NamedLogging
     with Spanning {
 
-  private def getStartTimestamp(implicit
+  protected def readLatestProcessedSnapshotTimestamp(implicit
       tc: TraceContext
-  ): Future[Option[TimestampWithMigrationId]] =
-    kvProvider.getLatestAcsSnapshotInBulkStorage().value
+  ): Future[Option[TimestampWithMigrationId]]
+
+  protected def persistLatestProcessedSnapshotTimestamp(ts: TimestampWithMigrationId)(implicit
+      tc: TraceContext
+  ): Future[Unit]
+
+  protected def getNextSnapshotTimestampAfter(
+      last: TimestampWithMigrationId
+  )(implicit tc: TraceContext): Future[Option[TimestampWithMigrationId]]
+
+  protected def shouldProcessSnapshotAt(ts: TimestampWithMigrationId)(implicit
+      tc: TraceContext
+  ): Boolean
+
+  protected def processSnapshotAt(ts: TimestampWithMigrationId)(implicit
+      tc: TraceContext
+  ): Flow[TimestampWithMigrationId, TimestampWithMigrationId, NotUsed]
+
+  protected def updateMetric(ts: TimestampWithMigrationId): Unit
 
   private def getAcsSnapshotTimestampsAfter(
       start: TimestampWithMigrationId
   )(implicit tc: TraceContext): Source[TimestampWithMigrationId, NotUsed] = {
     Source
       .unfoldAsync(start) { (last: TimestampWithMigrationId) =>
-        acsSnapshotStore.lookupSnapshotAfter(last.migrationId, last.timestamp).flatMap {
+        getNextSnapshotTimestampAfter(last).flatMap {
           case Some(snapshot) =>
             logger.info(
-              s"next snapshot available, at migration ${snapshot.migrationId}, record time ${snapshot.snapshotRecordTime}"
+              s"next snapshot available, at migration ${snapshot.migrationId}, record time ${snapshot.timestamp}"
             )
             Future.successful(
               Some(
                 (
-                  TimestampWithMigrationId(snapshot.snapshotRecordTime, snapshot.migrationId),
-                  Some(TimestampWithMigrationId(snapshot.snapshotRecordTime, snapshot.migrationId)),
+                  snapshot,
+                  Some(snapshot),
                 )
               )
             )
@@ -102,7 +106,7 @@ class AcsSnapshotBulkStorage(
 
     backfillingCompleteGate.flatMap { _ =>
       Source
-        .future(getStartTimestamp)
+        .future(readLatestProcessedSnapshotTimestamp)
         .flatMapConcat {
           case Some(start: TimestampWithMigrationId) =>
             logger.info(
@@ -113,48 +117,15 @@ class AcsSnapshotBulkStorage(
             logger.info("No dumped snapshots yet, starting from genesis")
             getAcsSnapshotTimestampsAfter(TimestampWithMigrationId(CantonTimestamp.MinValue, 0))
         }
-        .filter { ts =>
-          val ret = storageConfig.shouldDumpSnapshotToBulkStorage(ts.timestamp)
-          if (ret) {
-            logger.debug(s"Dumping snapshot at timestamp ${ts.timestamp} to bulk storage")
-          } else {
-            logger.info(
-              s"Skipping snapshot at timestamp ${ts.timestamp} for bulk storage, not required per the configured period of ${storageConfig.bulkAcsSnapshotPeriodHours}"
-            )
-          }
-          ret
-        }
+        .filter { shouldProcessSnapshotAt }
         .flatMapConcat(ts =>
           Source
             .single(ts)
-            .via(
-              SingleAcsSnapshotBulkStorage
-                .asFlow(
-                  storageConfig,
-                  appConfig,
-                  acsSnapshotStore,
-                  s3Connection,
-                  historyMetrics,
-                  loggerFactory,
-                )
-                .map(keys => {
-                  logger.debug(
-                    s"Successfully dumped snapshot from migration ${ts.migrationId}, timestamp ${ts.timestamp} to bulk storage, with object keys: $keys"
-                  )
-                  ts
-                })
-            )
+            .via(processSnapshotAt(ts))
         )
         .mapAsync(1) { ts =>
-          historyMetrics.BulkStorage.latestAcsSnapshot.updateValue(ts.timestamp)
-          kvProvider
-            .setLatestAcsSnapshotsInBulkStorage(ts)
-            .map(_ => {
-              logger.info(
-                s"Successfully completed dumping snapshots from migration ${ts.migrationId}, timestamp ${ts.timestamp}"
-              )
-              ts
-            })
+          updateMetric(ts)
+          persistLatestProcessedSnapshotTimestamp(ts).map(_ => ts)
         }
     }
   }
@@ -177,53 +148,6 @@ class AcsSnapshotBulkStorage(
       )
     }
   }
-
-  def getAcsSnapshotAtOrBefore(
-      atOrBeforeTimestamp: CantonTimestamp
-  )(implicit tc: TraceContext): Future[AcsSnapshotObjects] = {
-
-    for {
-      snapshotTs <- kvProvider
-        .getLatestAcsSnapshotInBulkStorage()
-        .value
-        .map {
-          case None =>
-            throw Status.NOT_FOUND
-              .withDescription("no snapshot in bulk storage yet")
-              .asRuntimeException()
-          case Some(ts) if ts.timestamp < atOrBeforeTimestamp =>
-            logger.trace(
-              s"Latest snapshot in bulk storage is at ${ts.timestamp}, which is before the requested timestamp ${atOrBeforeTimestamp}, returning that one"
-            )
-            ts.timestamp
-          case Some(ts) => storageConfig.computeBulkSnapshotTimeAtOrBefore(atOrBeforeTimestamp)
-        }
-      prefix = storageConfig.findSegmentFolderPrefixByStartTimestamp(snapshotTs)
-      objects <- s3Connection
-        // A single object currently holds ~700K contracts, we apply a Limit just for safety,
-        // but we don't expect to get anywhere near 1000 such objects in the foreseeable future
-        // (hence the HardLimit, just as a safety precaution).
-        .listObjects(
-          prefix,
-          _.matches(".*ACS_\\d+\\.zstd"),
-          HardLimit.tryCreate(Limit.DefaultMaxPageSize),
-        )
-      objectsWithChecksums <- s3Connection.getChecksums(objects)
-    } yield {
-      if (objects.isEmpty) {
-        throw Status.NOT_FOUND
-          .withDescription(
-            s"No snapshot objects found in bulk storage at expected timestamp at or before $atOrBeforeTimestamp, this may be because the timestamp is before network genesis"
-          )
-          .asRuntimeException()
-      }
-      logger.trace(
-        s"Found snapshot in bulk storage at timestamp $snapshotTs, with objects: ${objects.mkString(", ")}"
-      )
-      AcsSnapshotObjects(snapshotTs, objectsWithChecksums)
-    }
-  }
-
 }
 
 object AcsSnapshotBulkStorage {
