+++ apps/scan/src/test/scala/org/lfdecentralizedtrust/splice/scan/store/bulk/AcsSnapshotBulkStorageTest.scala
@@ -166,7 +166,7 @@ class AcsSnapshotBulkStorageTest
 
       val retryProvider =
         RetryProvider(loggerFactory, timeouts, FutureSupervisor.Noop, NoOpMetricsFactory)
-      val bulkStorage = new AcsSnapshotBulkStorage(
+      val bulkStorage = new AcsSnapshotBulkStorageStaging(
         bulkStorageTestConfig,
         appConfig,
         store.store,
